revert kubelet api to get the envvars
